+++ sdk/monitor/azquery/utils_test.go
@@ -30,13 +30,17 @@ const recordingDirectory = "sdk/monitor/azquery/testdata"
 const fakeWorkspaceID = "32d1e136-gg81-4b0a-b647-260cdc471f68"
 const fakeWorkspaceID2 = "asdjkfj8k20-gg81-4b0a-9fu2-260c09fn1f68"
 const fakeResourceURI = "/subscriptions/faa080af-c1d8-40ad-9cce-e1a451va7b87/resourceGroups/rg-example/providers/Microsoft.AppConfiguration/configurationStores/example"
+const fakeSubscrtiptionID = "faa080af-c1d8-40ad-9cce-e1a451va7b87"
+const fakeRegion = "westus2"
 
 var (
-	credential   azcore.TokenCredential
-	workspaceID  string
-	workspaceID2 string
-	resourceURI  string
-	clientCloud  cloud.Configuration
+	credential     azcore.TokenCredential
+	workspaceID    string
+	workspaceID2   string
+	resourceURI    string
+	subscriptionID string
+	region         string
+	clientCloud    cloud.Configuration
 )
 
 func TestMain(m *testing.M) {
@@ -83,6 +87,8 @@ func run(m *testing.M) int {
 	workspaceID = getEnvVar("WORKSPACE_ID", fakeWorkspaceID)
 	workspaceID2 = getEnvVar("WORKSPACE_ID2", fakeWorkspaceID2)
 	resourceURI = getEnvVar("RESOURCE_URI", fakeResourceURI)
+	subscriptionID = getEnvVar("AZQUERY_SUBSCRIPTION_ID", fakeSubscrtiptionID)
+	region = getEnvVar("AZQUERY_LOCATION", fakeRegion)
 
 	return m.Run()
 }
@@ -127,6 +133,19 @@ func startMetricsTest(t *testing.T) *azquery.MetricsClient {
 	}
 
 	client, err := azquery.NewMetricsClient(credential, opts)
+	if err != nil {
+		panic(err)
+	}
+	return client
+}
+
+func startMetricsBatchTest(t *testing.T) *azquery.MetricsBatchClient {
+	startRecording(t)
+	transport, err := recording.NewRecordingHTTPClient(t, nil)
+	require.NoError(t, err)
+	opts := &azquery.MetricsBatchClientOptions{ClientOptions: azcore.ClientOptions{Transport: transport}}
+	endpoint := "https://" + region + ".metrics.monitor.azure.com"
+	client, err := azquery.NewMetricsBatchClient(endpoint, credential, opts)
 	if err != nil {
 		panic(err)
 	}